chore: move `util::{copy, force_cononicalize_dir}` to `cargo_near_build::fs`
