+++ cargo-near-build/Cargo.toml
@@ -7,3 +7,4 @@ version = "0.1.0"
 eyre = "0.6.12"
 camino = "1.1.1"
 colored = "2.0"
+dunce = "1"
