+++ cargo-near-build/src/lib.rs
@@ -1,2 +1,5 @@
+// TODO: make mod non-pub
+pub mod fs;
+// TODO: make mod non-pub
 pub mod pretty_print;
 pub mod types;
